Fixes #107 by adding app re-exports
