+++ addon/instance-initializers/violations-helper.js
@@ -1,0 +1,11 @@
+import violationsHelper from 'ember-a11y-testing/utils/violations-helper';
+
+export function initialize() {
+  window.violationsHelper = violationsHelper;
+}
+
+export default {
+  name: 'violations-helper',
+  before: 'axe-component',
+  initialize
+};
